Create a util file for "getExistingReferralId" function
Related to #118
